Add ZDEMO command that, when the player saves a game, records its name.
This is used to replay the last savegame name without having to actually
run the whole log through the ZZT engine.

There's a known bug where the demo may err with error 35, but I haven't
tracked down what the cause is yet. It might be something with the PIT
counter.