Add stub CGI processing scripts.
These will be turned into verification scripts later. The following
needs to be implemented:

get.py:
	- Synthesize .HI files given database (user, high score)
		records.

put.py:
	- Accept .ZZD uploads and store in data directory if not already
		stored.
	- Parse .ZZD format to check for validity.
	- Hand off to native zeta ZDEMOVAL to verify score.
	- Insert score into database.

put.py works by posting a JSON dictionary with keys "filename" and "value",
the former being a full filename (including extension) and the latter being
a base 64 encoding of the file to be uploaded.

get.py accepts the file name as a HTTP parameter with key "filename", and
should return the synthesized high score file as a JSON with the same
format as what put.py takes.